backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
flask-bcrypt
flask-jwt-extended
python-dotenv
apscheduler
requests
psycopg2-binary
edge-tts
